+++ src/Npgsql/NpgsqlDataReaderBase.cs
@@ -1,0 +1,436 @@
+﻿using System;
+using System.Collections;
+using System.Collections.Generic;
+using System.Collections.ObjectModel;
+using System.Data;
+using System.Data.Common;
+using System.IO;
+using System.Linq;
+using System.Threading;
+using System.Threading.Tasks;
+using JetBrains.Annotations;
+using Npgsql.PostgresTypes;
+using Npgsql.Schema;
+using NpgsqlTypes;
+
+namespace Npgsql
+{
+    /// <summary>
+    /// Reads a forward-only stream of rows from a data source.
+    /// </summary>
+    /// <remarks>
+    /// Useful shared overrides of <see cref="DbDataReader"/> along with the additional public and internal
+    /// API contracts of <see cref="NpgsqlDataReader"/>.
+    /// NOTE: Possibly everything public abstract in here could be marked with <see cref="PublicAPIAttribute"/>?
+    /// </remarks>
+    public abstract class NpgsqlDataReader : DbDataReader
+#if !NET461
+        , IDbColumnSchemaGenerator
+#endif
+    {
+        internal NpgsqlCommand Command { get; private protected set; } = default!;
+        private protected readonly NpgsqlConnector Connector;
+
+        /// <summary>
+        /// Is raised whenever Close() is called.
+        /// </summary>
+        public abstract event EventHandler? ReaderClosed;
+
+        /// <summary>
+        /// Gets a value indicating the depth of nesting for the current row.  Always returns zero.
+        /// </summary>
+        public override int Depth => 0;
+
+        /// <summary>
+        /// Returns details about each statement that this reader will or has executed.
+        /// </summary>
+        /// <remarks>
+        /// Note that some fields (i.e. rows and oid) are only populated as the reader
+        /// traverses the result.
+        ///
+        /// For commands with multiple queries, this exposes the number of rows affected on
+        /// a statement-by-statement basis, unlike <see cref="DbDataReader.RecordsAffected"/>
+        /// which exposes an aggregation across all statements.
+        /// </remarks>
+        public abstract IReadOnlyList<NpgsqlStatement> Statements { get; }
+
+        /// <summary>
+        /// Indicates whether the reader is currently positioned on a row, i.e. whether reading a
+        /// column is possible.
+        /// This property is different from <see cref="DbDataReader.HasRows"/> in that <see cref="DbDataReader.HasRows"/> will
+        /// return true even if attempting to read a column will fail, e.g. before <see cref="DbDataReader.Read()"/>
+        /// has been called
+        /// </summary>
+        [PublicAPI]
+        public abstract bool IsOnRow { get; }
+
+        internal NpgsqlDataReader(NpgsqlConnector connector)
+        {
+            Connector = connector;
+        }
+
+        #region Cleanup / Dispose
+
+        /// <summary>
+        /// Releases the resources used by the <see cref="NpgsqlDataReader">NpgsqlDataReader</see>.
+        /// </summary>
+        protected override void Dispose(bool disposing) => Close();
+
+        /// <summary>
+        /// Releases the resources used by the <see cref="NpgsqlDataReader">NpgsqlDataReader</see>.
+        /// </summary>
+#if !NET461 && !NETSTANDARD2_0
+        public override ValueTask DisposeAsync()
+#else
+        public ValueTask DisposeAsync()
+#endif
+        {
+            using (NoSynchronizationContextScope.Enter())
+                return new ValueTask(Close(connectionClosing: false, async: true));
+        }
+
+        /// <summary>
+        /// Closes the <see cref="NpgsqlDataReader"/> reader, allowing a new command to be executed.
+        /// </summary>
+        public override void Close() => Close(connectionClosing: false, async: false).GetAwaiter().GetResult();
+
+        /// <summary>
+        /// Closes the <see cref="NpgsqlDataReader"/> reader, allowing a new command to be executed.
+        /// </summary>
+#if !NET461 && !NETSTANDARD2_0
+        public override Task CloseAsync()
+#else
+        public Task CloseAsync()
+#endif
+            => Close(connectionClosing: false, async: true);
+
+        internal abstract Task Close(bool connectionClosing, bool async);
+
+        internal abstract Task Cleanup(bool async, bool connectionClosing = false);
+
+        #endregion
+
+        #region Simple value getters
+
+        /// <summary>
+        /// Gets the value of the specified column as a Boolean.
+        /// </summary>
+        /// <param name="ordinal">The zero-based column ordinal.</param>
+        /// <returns>The value of the specified column.</returns>
+        public override bool GetBoolean(int ordinal) => GetFieldValue<bool>(ordinal);
+
+        /// <summary>
+        /// Gets the value of the specified column as a byte.
+        /// </summary>
+        /// <param name="ordinal">The zero-based column ordinal.</param>
+        /// <returns>The value of the specified column.</returns>
+        public override byte GetByte(int ordinal) => GetFieldValue<byte>(ordinal);
+
+        /// <summary>
+        /// Gets the value of the specified column as a single character.
+        /// </summary>
+        /// <param name="ordinal">The zero-based column ordinal.</param>
+        /// <returns>The value of the specified column.</returns>
+        public override char GetChar(int ordinal) => GetFieldValue<char>(ordinal);
+
+        /// <summary>
+        /// Gets the value of the specified column as a 16-bit signed integer.
+        /// </summary>
+        /// <param name="ordinal">The zero-based column ordinal.</param>
+        /// <returns>The value of the specified column.</returns>
+        public override short GetInt16(int ordinal) => GetFieldValue<short>(ordinal);
+
+        /// <summary>
+        /// Gets the value of the specified column as a 32-bit signed integer.
+        /// </summary>
+        /// <param name="ordinal">The zero-based column ordinal.</param>
+        /// <returns>The value of the specified column.</returns>
+        public override int GetInt32(int ordinal) => GetFieldValue<int>(ordinal);
+
+        /// <summary>
+        /// Gets the value of the specified column as a 64-bit signed integer.
+        /// </summary>
+        /// <param name="ordinal">The zero-based column ordinal.</param>
+        /// <returns>The value of the specified column.</returns>
+        public override long GetInt64(int ordinal) => GetFieldValue<long>(ordinal);
+
+        /// <summary>
+        /// Gets the value of the specified column as a <see cref="DateTime"/> object.
+        /// </summary>
+        /// <param name="ordinal">The zero-based column ordinal.</param>
+        /// <returns>The value of the specified column.</returns>
+        public override DateTime GetDateTime(int ordinal) => GetFieldValue<DateTime>(ordinal);
+
+        /// <summary>
+        /// Gets the value of the specified column as an instance of <see cref="string"/>.
+        /// </summary>
+        /// <param name="ordinal">The zero-based column ordinal.</param>
+        /// <returns>The value of the specified column.</returns>
+        public override string GetString(int ordinal) => GetFieldValue<string>(ordinal);
+
+        /// <summary>
+        /// Gets the value of the specified column as a <see cref="decimal"/> object.
+        /// </summary>
+        /// <param name="ordinal">The zero-based column ordinal.</param>
+        /// <returns>The value of the specified column.</returns>
+        public override decimal GetDecimal(int ordinal) => GetFieldValue<decimal>(ordinal);
+
+        /// <summary>
+        /// Gets the value of the specified column as a double-precision floating point number.
+        /// </summary>
+        /// <param name="ordinal">The zero-based column ordinal.</param>
+        /// <returns>The value of the specified column.</returns>
+        public override double GetDouble(int ordinal) => GetFieldValue<double>(ordinal);
+
+        /// <summary>
+        /// Gets the value of the specified column as a single-precision floating point number.
+        /// </summary>
+        /// <param name="ordinal">The zero-based column ordinal.</param>
+        /// <returns>The value of the specified column.</returns>
+        public override float GetFloat(int ordinal) => GetFieldValue<float>(ordinal);
+
+        /// <summary>
+        /// Gets the value of the specified column as a globally-unique identifier (GUID).
+        /// </summary>
+        /// <param name="ordinal">The zero-based column ordinal.</param>
+        /// <returns>The value of the specified column.</returns>
+        public override Guid GetGuid(int ordinal) => GetFieldValue<Guid>(ordinal);
+
+        /// <summary>
+        /// Gets the value of the specified column as an instance of <see cref="object"/>.
+        /// </summary>
+        /// <param name="ordinal">The zero-based column ordinal.</param>
+        /// <returns>The value of the specified column.</returns>
+        public override object this[int ordinal] => GetValue(ordinal);
+
+        #endregion
+
+        #region Provider-specific simple type getters
+
+        /// <summary>
+        /// Gets the value of the specified column as an <see cref="NpgsqlDate"/>,
+        /// Npgsql's provider-specific type for dates.
+        /// </summary>
+        /// <remarks>
+        /// PostgreSQL's date type represents dates from 4713 BC to 5874897 AD, while .NET's DateTime
+        /// only supports years from 1 to 1999. If you require years outside this range use this accessor.
+        /// The standard <see cref="DbDataReader.GetProviderSpecificValue"/> method will also return this type, but has
+        /// the disadvantage of boxing the value.
+        /// See http://www.postgresql.org/docs/current/static/datatype-datetime.html
+        /// </remarks>
+        /// <param name="ordinal">The zero-based column ordinal.</param>
+        /// <returns>The value of the specified column.</returns>
+        public NpgsqlDate GetDate(int ordinal) => GetFieldValue<NpgsqlDate>(ordinal);
+
+        /// <summary>
+        /// Gets the value of the specified column as a TimeSpan,
+        /// </summary>
+        /// <remarks>
+        /// PostgreSQL's interval type has has a resolution of 1 microsecond and ranges from
+        /// -178000000 to 178000000 years, while .NET's TimeSpan has a resolution of 100 nanoseconds
+        /// and ranges from roughly -29247 to 29247 years.
+        /// See http://www.postgresql.org/docs/current/static/datatype-datetime.html
+        /// </remarks>
+        /// <param name="ordinal">The zero-based column ordinal.</param>
+        /// <returns>The value of the specified column.</returns>
+        public TimeSpan GetTimeSpan(int ordinal) => GetFieldValue<TimeSpan>(ordinal);
+
+        /// <summary>
+        /// Gets the value of the specified column as an <see cref="NpgsqlTimeSpan"/>,
+        /// Npgsql's provider-specific type for time spans.
+        /// </summary>
+        /// <remarks>
+        /// PostgreSQL's interval type has has a resolution of 1 microsecond and ranges from
+        /// -178000000 to 178000000 years, while .NET's TimeSpan has a resolution of 100 nanoseconds
+        /// and ranges from roughly -29247 to 29247 years. If you require values from outside TimeSpan's
+        /// range use this accessor.
+        /// The standard ADO.NET <see cref="DbDataReader.GetProviderSpecificValue"/> method will also return this
+        /// type, but has the disadvantage of boxing the value.
+        /// See http://www.postgresql.org/docs/current/static/datatype-datetime.html
+        /// </remarks>
+        /// <param name="ordinal">The zero-based column ordinal.</param>
+        /// <returns>The value of the specified column.</returns>
+        public NpgsqlTimeSpan GetInterval(int ordinal) => GetFieldValue<NpgsqlTimeSpan>(ordinal);
+
+        /// <summary>
+        /// Gets the value of the specified column as an <see cref="NpgsqlDateTime"/>,
+        /// Npgsql's provider-specific type for date/time timestamps. Note that this type covers
+        /// both PostgreSQL's "timestamp with time zone" and "timestamp without time zone" types,
+        /// which differ only in how they are converted upon input/output.
+        /// </summary>
+        /// <remarks>
+        /// PostgreSQL's timestamp type represents dates from 4713 BC to 5874897 AD, while .NET's DateTime
+        /// only supports years from 1 to 1999. If you require years outside this range use this accessor.
+        /// The standard <see cref="DbDataReader.GetProviderSpecificValue"/> method will also return this type, but has
+        /// the disadvantage of boxing the value.
+        /// See http://www.postgresql.org/docs/current/static/datatype-datetime.html
+        /// </remarks>
+        /// <param name="ordinal">The zero-based column ordinal.</param>
+        /// <returns>The value of the specified column.</returns>
+        public NpgsqlDateTime GetTimeStamp(int ordinal) => GetFieldValue<NpgsqlDateTime>(ordinal);
+
+        #endregion
+
+        #region Special binary getters
+
+        /// <summary>
+        /// Retrieves data as a <see cref="Stream"/>.
+        /// </summary>
+        /// <param name="ordinal">The zero-based column ordinal.</param>
+        /// <param name="cancellationToken">The token to monitor for cancellation requests. The default value is <see cref="CancellationToken.None"/>.</param>
+        /// <returns>The returned object.</returns>
+        public abstract Task<Stream> GetStreamAsync(int ordinal, CancellationToken cancellationToken = default);
+
+        #endregion
+
+        #region Special text getters
+
+        /// <summary>
+        /// Retrieves data as a <see cref="TextReader"/>.
+        /// </summary>
+        /// <param name="ordinal">The zero-based column ordinal.</param>
+        /// <param name="cancellationToken">The token to monitor for cancellation requests. The default value is <see cref="CancellationToken.None"/>.</param>
+        /// <returns>The returned object.</returns>
+        public abstract Task<TextReader> GetTextReaderAsync(int ordinal, CancellationToken cancellationToken = default);
+
+        #endregion
+
+        #region GetValue
+
+        /// <summary>
+        /// Gets the value of the specified column as an instance of <see cref="object"/>.
+        /// </summary>
+        /// <param name="name">The name of the column.</param>
+        /// <returns>The value of the specified column.</returns>
+        public override object this[string name] => GetValue(GetOrdinal(name));
+
+        #endregion
+
+        #region Other public accessors
+
+        /// <summary>
+        /// Gets a representation of the PostgreSQL data type for the specified field.
+        /// The returned representation can be used to access various information about the field.
+        /// </summary>
+        /// <param name="ordinal">The zero-based column index.</param>
+        [PublicAPI]
+        public abstract PostgresType GetPostgresType(int ordinal);
+
+        /// <summary>
+        /// Gets the OID for the PostgreSQL type for the specified field, as it appears in the pg_type table.
+        /// </summary>
+        /// <remarks>
+        /// This is a PostgreSQL-internal value that should not be relied upon and should only be used for
+        /// debugging purposes.
+        /// </remarks>
+        /// <param name="ordinal">The zero-based column index.</param>
+        public abstract uint GetDataTypeOID(int ordinal);
+
+        /// <summary>
+        /// Returns an <see cref="IEnumerator"/> that can be used to iterate through the rows in the data reader.
+        /// </summary>
+        /// <returns>An <see cref="IEnumerator"/> that can be used to iterate through the rows in the data reader.</returns>
+        public override IEnumerator GetEnumerator()
+            => new DbEnumerator(this);
+
+        /// <summary>
+        /// Returns schema information for the columns in the current resultset.
+        /// </summary>
+        /// <returns></returns>
+        public abstract ReadOnlyCollection<NpgsqlDbColumn> GetColumnSchema();
+
+#if !NET461
+        ReadOnlyCollection<DbColumn> IDbColumnSchemaGenerator.GetColumnSchema()
+            => new ReadOnlyCollection<DbColumn>(GetColumnSchema().Cast<DbColumn>().ToList());
+#endif
+
+        #endregion
+
+        #region Schema metadata table
+
+        /// <summary>
+        /// Returns a System.Data.DataTable that describes the column metadata of the DataReader.
+        /// </summary>
+#nullable disable
+        public override DataTable GetSchemaTable()
+#nullable restore
+        {
+            if (FieldCount == 0) // No resultset
+                return null;
+
+            var table = new DataTable("SchemaTable");
+
+            // Note: column order is important to match SqlClient's, some ADO.NET users appear
+            // to assume ordering (see #1671)
+            table.Columns.Add("ColumnName", typeof(string));
+            table.Columns.Add("ColumnOrdinal", typeof(int));
+            table.Columns.Add("ColumnSize", typeof(int));
+            table.Columns.Add("NumericPrecision", typeof(int));
+            table.Columns.Add("NumericScale", typeof(int));
+            table.Columns.Add("IsUnique", typeof(bool));
+            table.Columns.Add("IsKey", typeof(bool));
+            table.Columns.Add("BaseServerName", typeof(string));
+            table.Columns.Add("BaseCatalogName", typeof(string));
+            table.Columns.Add("BaseColumnName", typeof(string));
+            table.Columns.Add("BaseSchemaName", typeof(string));
+            table.Columns.Add("BaseTableName", typeof(string));
+            table.Columns.Add("DataType", typeof(Type));
+            table.Columns.Add("AllowDBNull", typeof(bool));
+            table.Columns.Add("ProviderType", typeof(int));
+            table.Columns.Add("IsAliased", typeof(bool));
+            table.Columns.Add("IsExpression", typeof(bool));
+            table.Columns.Add("IsIdentity", typeof(bool));
+            table.Columns.Add("IsAutoIncrement", typeof(bool));
+            table.Columns.Add("IsRowVersion", typeof(bool));
+            table.Columns.Add("IsHidden", typeof(bool));
+            table.Columns.Add("IsLong", typeof(bool));
+            table.Columns.Add("IsReadOnly", typeof(bool));
+            table.Columns.Add("ProviderSpecificDataType", typeof(Type));
+            table.Columns.Add("DataTypeName", typeof(string));
+
+            foreach (var column in GetColumnSchema())
+            {
+                var row = table.NewRow();
+
+                row["ColumnName"] = column.ColumnName;
+                row["ColumnOrdinal"] = column.ColumnOrdinal ?? -1;
+                row["ColumnSize"] = column.ColumnSize ?? -1;
+                row["NumericPrecision"] = column.NumericPrecision ?? 0;
+                row["NumericScale"] = column.NumericScale ?? 0;
+                row["IsUnique"] = column.IsUnique == true;
+                row["IsKey"] = column.IsKey == true;
+                row["BaseServerName"] = "";
+                row["BaseCatalogName"] = column.BaseCatalogName;
+                row["BaseColumnName"] = column.BaseColumnName;
+                row["BaseSchemaName"] = column.BaseSchemaName;
+                row["BaseTableName"] = column.BaseTableName;
+                row["DataType"] = column.DataType;
+                row["AllowDBNull"] = (object?)column.AllowDBNull ?? DBNull.Value;
+                row["ProviderType"] = column.NpgsqlDbType ?? NpgsqlDbType.Unknown;
+                row["IsAliased"] = column.IsAliased == true;
+                row["IsExpression"] = column.IsExpression == true;
+                row["IsIdentity"] = column.IsIdentity == true;
+                row["IsAutoIncrement"] = column.IsAutoIncrement == true;
+                row["IsRowVersion"] = false;
+                row["IsHidden"] = column.IsHidden == true;
+                row["IsLong"] = column.IsLong == true;
+                row["DataTypeName"] = column.DataTypeName;
+
+                table.Rows.Add(row);
+            }
+
+            return table;
+        }
+
+        #endregion Schema metadata table
+    }
+
+    enum ReaderState
+    {
+        BeforeResult,
+        InResult,
+        BetweenResults,
+        Consumed,
+        Closed,
+    }
+}
